Extract Committer trait
